Add generic enum value conversion

+++ src/AppGui.cpp
@@ -53,7 +53,7 @@ AppGui::AppGui(int & argc, char ** argv) :
 
 bool AppGui::initialize()
 {
-    qsrand(static_cast<uint>(time(nullptr)));
+    qsrand(time(nullptr));
 
     QCoreApplication::setOrganizationName("mooltipass");
     QCoreApplication::setOrganizationDomain("themooltipass.com");
